+++ microwave/util/inc/data_types.h
@@ -1,10 +1,10 @@
 #ifndef _DATA_TYPES_H_
 #define _DATA_TYPES_H_
 
-typedef unsigned char   uint8;
-typedef unsigned short  uint16;
-typedef unsigned int    uint32;
-typedef unsigned long   uint64;
+typedef unsigned char       uint8;
+typedef unsigned int        uint16;
+typedef unsigned long       uint32;
+typedef unsigned long long  uint64;
 
 typedef enum {
     IntEn_TMR1_CCP2,
